+++ src/Form.php
@@ -25,6 +25,11 @@
     protected $validationResult = null;
 
     /**
+     * @var array
+     */
+    protected $errorList = [];
+
+    /**
      * @var array|null
      */
     protected $data = null;
@@ -154,6 +159,7 @@
      *
      */
     protected function cleanValidationFlag() {
+      $this->errorList = [];
       $this->validationResult = null;
     }
 
@@ -175,11 +181,34 @@
       $this->validationResult = true;
       foreach ($this->elements as $element) {
         if (!$element->isValid()) {
+          $this->errorList = array_merge($this->errorList, $element->getValidatorsErrors());
           $this->validationResult = false;
         }
       }
 
       return $this->validationResult;
+    }
+
+
+    /**
+     * @return array
+     */
+    public function getErrors() {
+      return $this->errorList;
+    }
+
+
+    /**
+     * @param string $error
+     * @return $this
+     */
+    protected function addError($error) {
+      if (!is_string($error)) {
+        throw new \InvalidArgumentException('Error should be a string, ' . gettype($error) . ' given.');
+      }
+      $this->validationResult = false;
+      $this->errorList[] = $error;
+      return $this;
     }
 
 
@@ -212,6 +241,19 @@
      */
     public function getElements() {
       return $this->elements;
+    }
+
+
+    /**
+     * @param string $name
+     * @return Element\BaseElement
+     * @throws \InvalidArgumentException
+     */
+    public function getElement($name) {
+      if (empty($this->elements[$name])) {
+        throw new \InvalidArgumentException('Element with name "' . $name . '" not found');
+      }
+      return $this->elements[$name];
     }
 
 
@@ -451,5 +493,4 @@
     public function renderEnd() {
       return '</form>';
     }
-
   }